backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
pymongo==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
reportlab==4.2.2
boto3==1.34.162
